requirements.txt
fastapi
uvicorn
scikit-learn
numpy
httpx
python-dotenv